feat: Moved to CMake

+++ tests/runtest.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
 
+if [ $# != 1 ]; then
+  echo >&2 "Expected testing app file name!\nAborting."
+  exit 1
+fi
+
 set -e
 PREFIX="`hostname`-`id --user`-`echo $$`-"
 IMAGE_NAME="${PREFIX}superiormysqlpp-test-mysql"
 CONTAINER_NAME="${PREFIX}superiormysqlpp-testdb"
-docker build --pull --tag=${IMAGE_NAME} ../db
+docker build --pull --tag=${IMAGE_NAME} db
 docker rm --force ${CONTAINER_NAME} >/dev/null 2>&1 || true
-docker run --detach --publish-all --name ${CONTAINER_NAME} ${IMAGE_NAME} #1>/dev/null
+docker run --detach --publish-all --name ${CONTAINER_NAME} ${IMAGE_NAME}
 MYSQL_HOST=`docker inspect --format='{{.NetworkSettings.IPAddress}}' ${CONTAINER_NAME}`
 set +e
 
@@ -16,7 +21,7 @@ LOCAL_TMP_SOCKET="/tmp/${CONTAINER_NAME}.sock"
 socat "UNIX-LISTEN:$LOCAL_TMP_SOCKET,fork,reuseaddr,unlink-early,mode=777" \
     TCP:$MYSQL_HOST:3307 &
 
-./tester ${MYSQL_HOST} 3306 ${CONTAINER_NAME} ${LOCAL_TMP_SOCKET} --reporter=spec
+$1 ${MYSQL_HOST} 3306 ${CONTAINER_NAME} ${LOCAL_TMP_SOCKET} --reporter=spec
 RESULT=$?
 
 ## TODO Add option to run test in manual mode.
